changed population to float & standardized debug print statments to print float values for pop, food, minerals, etc

+++ tiles/Empire.hpp
@@ -14,7 +14,7 @@ class Empire {
 
     public: 
         Empire();
-        Empire(int food, int population,  int number_of_towns, int minerals, std::string name);
+        Empire(float food, float population,  int number_of_towns, float minerals, std::string name);
         ~Empire();
 
         bool isEmpireGrowing();
@@ -27,10 +27,10 @@ class Empire {
         void printEmpire();
 
     private: 
-        int food;               /* Used to calculate growth */
-        int population;         /* Used to calculate growth */
+        float food;               /* Used to calculate growth */
+        float population;         /* Used to calculate growth */
         int number_of_towns;    /* Used for influence multiplier */
-        int minerals;           /* Used for military strength */
+        float minerals;           /* Used for military strength */
         float influence;
         std::string name;
